Integrate nested struct handling
