+++ datafusion/datasource/src/schema_adapter.rs
@@ -22,9 +22,10 @@
 //! can be stored external to a parquet file that maps parquet logical types to arrow types.
 use arrow::{
     array::{new_null_array, ArrayRef, RecordBatch, RecordBatchOptions},
-    compute::{can_cast_types, cast},
-    datatypes::{Field, Schema, SchemaRef},
+    compute::can_cast_types,
+    datatypes::{DataType::Struct, Field, Schema, SchemaRef},
 };
+use datafusion_common::nested_struct::adapt_column;
 use datafusion_common::{plan_err, ColumnStatistics};
 use std::{fmt::Debug, sync::Arc};
 /// Function used by [`SchemaMapping`] to adapt a column from the file schema to
@@ -274,16 +275,20 @@ impl SchemaAdapter for DefaultSchemaAdapter {
         let (field_mappings, projection) = create_field_mapping(
             file_schema,
             &self.projected_table_schema,
-            can_cast_field,
+            |file_field, table_field| match (
+                file_field.data_type(),
+                table_field.data_type(),
+            ) {
+                (Struct(_), Struct(_)) => Ok(true),
+                _ => can_cast_field(file_field, table_field),
+            },
         )?;
 
         Ok((
             Arc::new(SchemaMapping::new(
                 Arc::clone(&self.projected_table_schema),
                 field_mappings,
-                Arc::new(|array: &ArrayRef, field: &Field| {
-                    Ok(cast(array, field.data_type())?)
-                }),
+                Arc::new(|array: &ArrayRef, field: &Field| adapt_column(array, field)),
             )),
             projection,
         ))
@@ -441,14 +446,19 @@ impl SchemaMapper for SchemaMapping {
 }
 
 #[cfg(test)]
+#[allow(unused_qualifications)]
 mod tests {
     use super::*;
     use arrow::{
-        array::ArrayRef,
+        array::{Array, ArrayRef, StringBuilder, StructArray, TimestampMillisecondArray},
         compute::cast,
-        datatypes::{DataType, Field},
+        datatypes::{DataType, Field, SchemaRef, TimeUnit},
+        record_batch::RecordBatch,
     };
-    use datafusion_common::{stats::Precision, Statistics};
+    use datafusion_common::{
+        stats::Precision, ColumnStatistics, Result, ScalarValue, Statistics,
+    };
+    use std::sync::Arc;
 
     #[test]
     fn test_schema_mapping_map_statistics_basic() {
@@ -728,5 +738,457 @@ mod tests {
         let c_array = mapped_batch.column(2);
         assert_eq!(c_array.len(), 2);
         assert_eq!(c_array.null_count(), 2);
+    }
+
+    // ================================
+    // Nested Struct Adaptation Tests
+    // ================================
+
+    /// Helper function to create a basic nested schema with additionalInfo
+    fn create_basic_nested_schema() -> SchemaRef {
+        Arc::new(Schema::new(vec![create_additional_info_field(false)]))
+    }
+
+    /// Helper function to create a deeply nested schema with additionalInfo including reason field
+    fn create_deep_nested_schema() -> SchemaRef {
+        Arc::new(Schema::new(vec![create_additional_info_field(true)]))
+    }
+
+    /// Helper function to create the additionalInfo field with or without the reason subfield
+    fn create_additional_info_field(with_reason: bool) -> Field {
+        let mut field_children = vec![
+            Field::new("location", DataType::Utf8, true),
+            Field::new(
+                "timestamp_utc",
+                DataType::Timestamp(TimeUnit::Millisecond, None),
+                true,
+            ),
+        ];
+
+        if with_reason {
+            field_children.push(create_reason_field());
+        }
+
+        Field::new(
+            "additionalInfo",
+            DataType::Struct(field_children.into()),
+            true,
+        )
+    }
+
+    /// Helper function to create the reason nested field with its details subfield
+    fn create_reason_field() -> Field {
+        Field::new(
+            "reason",
+            DataType::Struct(
+                vec![
+                    Field::new("_level", DataType::Float64, true),
+                    Field::new(
+                        "details",
+                        DataType::Struct(
+                            vec![
+                                Field::new("rurl", DataType::Utf8, true),
+                                Field::new("s", DataType::Float64, true),
+                                Field::new("t", DataType::Utf8, true),
+                            ]
+                            .into(),
+                        ),
+                        true,
+                    ),
+                ]
+                .into(),
+            ),
+            true,
+        )
+    }
+
+    #[test]
+    fn test_adapt_struct_with_added_nested_fields() -> Result<()> {
+        let (file_schema, table_schema) = create_test_schemas_with_nested_fields();
+
+        let batch = create_test_batch_with_struct_data(&file_schema)?;
+
+        let adapter = DefaultSchemaAdapterFactory
+            .create(Arc::clone(&table_schema), Arc::clone(&table_schema));
+        let (mapper, _) = adapter.map_schema(file_schema.as_ref())?;
+        let mapped_batch = mapper.map_batch(batch)?;
+
+        verify_adapted_batch_with_nested_fields(&mapped_batch, &table_schema)?;
+        Ok(())
+    }
+
+    fn create_test_schemas_with_nested_fields() -> (SchemaRef, SchemaRef) {
+        let file_schema = Arc::new(Schema::new(vec![Field::new(
+            "info",
+            DataType::Struct(
+                vec![
+                    Field::new("location", DataType::Utf8, true),
+                    Field::new(
+                        "timestamp_utc",
+                        DataType::Timestamp(TimeUnit::Millisecond, Some("UTC".into())),
+                        true,
+                    ),
+                ]
+                .into(),
+            ),
+            true,
+        )]));
+
+        let table_schema = Arc::new(Schema::new(vec![Field::new(
+            "info",
+            DataType::Struct(
+                vec![
+                    Field::new("location", DataType::Utf8, true),
+                    Field::new(
+                        "timestamp_utc",
+                        DataType::Timestamp(TimeUnit::Millisecond, Some("UTC".into())),
+                        true,
+                    ),
+                    Field::new(
+                        "reason",
+                        DataType::Struct(
+                            vec![
+                                Field::new("_level", DataType::Float64, true),
+                                Field::new(
+                                    "details",
+                                    DataType::Struct(
+                                        vec![
+                                            Field::new("rurl", DataType::Utf8, true),
+                                            Field::new("s", DataType::Float64, true),
+                                            Field::new("t", DataType::Utf8, true),
+                                        ]
+                                        .into(),
+                                    ),
+                                    true,
+                                ),
+                            ]
+                            .into(),
+                        ),
+                        true,
+                    ),
+                ]
+                .into(),
+            ),
+            true,
+        )]));
+
+        (file_schema, table_schema)
+    }
+
+    fn create_test_batch_with_struct_data(
+        file_schema: &SchemaRef,
+    ) -> Result<RecordBatch> {
+        let mut location_builder = StringBuilder::new();
+        location_builder.append_value("San Francisco");
+        location_builder.append_value("New York");
+
+        let timestamp_array = TimestampMillisecondArray::from(vec![
+            Some(1640995200000),
+            Some(1641081600000),
+        ]);
+
+        let timestamp_type =
+            DataType::Timestamp(TimeUnit::Millisecond, Some("UTC".into()));
+        let timestamp_array = cast(&timestamp_array, &timestamp_type)?;
+
+        let info_struct = StructArray::from(vec![
+            (
+                Arc::new(Field::new("location", DataType::Utf8, true)),
+                Arc::new(location_builder.finish()) as Arc<dyn Array>,
+            ),
+            (
+                Arc::new(Field::new("timestamp_utc", timestamp_type, true)),
+                timestamp_array,
+            ),
+        ]);
+
+        Ok(RecordBatch::try_new(
+            Arc::clone(file_schema),
+            vec![Arc::new(info_struct)],
+        )?)
+    }
+
+    fn verify_adapted_batch_with_nested_fields(
+        mapped_batch: &RecordBatch,
+        table_schema: &SchemaRef,
+    ) -> Result<()> {
+        assert_eq!(mapped_batch.schema(), *table_schema);
+        assert_eq!(mapped_batch.num_rows(), 2);
+
+        let info_col = mapped_batch.column(0);
+        let info_array = info_col
+            .as_any()
+            .downcast_ref::<StructArray>()
+            .expect("Expected info column to be a StructArray");
+
+        verify_preserved_fields(info_array)?;
+        verify_reason_field_structure(info_array)?;
+
+        Ok(())
+    }
+
+    fn verify_preserved_fields(info_array: &StructArray) -> Result<()> {
+        let location_col = info_array
+            .column_by_name("location")
+            .expect("Expected location field in struct");
+        let location_array = location_col
+            .as_any()
+            .downcast_ref::<arrow::array::StringArray>()
+            .expect("Expected location to be a StringArray");
+
+        assert_eq!(location_array.value(0), "San Francisco");
+        assert_eq!(location_array.value(1), "New York");
+
+        let timestamp_col = info_array
+            .column_by_name("timestamp_utc")
+            .expect("Expected timestamp_utc field in struct");
+        let timestamp_array = timestamp_col
+            .as_any()
+            .downcast_ref::<TimestampMillisecondArray>()
+            .expect("Expected timestamp_utc to be a TimestampMillisecondArray");
+
+        assert_eq!(timestamp_array.value(0), 1640995200000);
+        assert_eq!(timestamp_array.value(1), 1641081600000);
+
+        Ok(())
+    }
+
+    fn verify_reason_field_structure(info_array: &StructArray) -> Result<()> {
+        let reason_col = info_array
+            .column_by_name("reason")
+            .expect("Expected reason field in struct");
+        let reason_array = reason_col
+            .as_any()
+            .downcast_ref::<StructArray>()
+            .expect("Expected reason to be a StructArray");
+
+        assert_eq!(reason_array.fields().len(), 2);
+        assert!(reason_array.column_by_name("_level").is_some());
+        assert!(reason_array.column_by_name("details").is_some());
+
+        let details_col = reason_array
+            .column_by_name("details")
+            .expect("Expected details field in reason struct");
+        let details_array = details_col
+            .as_any()
+            .downcast_ref::<StructArray>()
+            .expect("Expected details to be a StructArray");
+
+        assert_eq!(details_array.fields().len(), 3);
+        assert!(details_array.column_by_name("rurl").is_some());
+        assert!(details_array.column_by_name("s").is_some());
+        assert!(details_array.column_by_name("t").is_some());
+
+        for i in 0..2 {
+            assert!(reason_array.is_null(i), "reason field should be null");
+        }
+
+        Ok(())
+    }
+
+    fn verify_column_statistics(
+        stats: &ColumnStatistics,
+        expected_null_count: Option<usize>,
+        expected_distinct_count: Option<usize>,
+        expected_min: Option<ScalarValue>,
+        expected_max: Option<ScalarValue>,
+        expected_sum: Option<ScalarValue>,
+    ) {
+        if let Some(count) = expected_null_count {
+            assert_eq!(
+                stats.null_count,
+                datafusion_common::stats::Precision::Exact(count),
+            );
+        }
+
+        if let Some(count) = expected_distinct_count {
+            assert_eq!(
+                stats.distinct_count,
+                datafusion_common::stats::Precision::Exact(count),
+            );
+        }
+
+        if let Some(min) = expected_min {
+            assert_eq!(
+                stats.min_value,
+                datafusion_common::stats::Precision::Exact(min),
+            );
+        }
+
+        if let Some(max) = expected_max {
+            assert_eq!(
+                stats.max_value,
+                datafusion_common::stats::Precision::Exact(max),
+            );
+        }
+
+        if let Some(sum) = expected_sum {
+            assert_eq!(
+                stats.sum_value,
+                datafusion_common::stats::Precision::Exact(sum),
+            );
+        }
+    }
+
+    fn create_test_column_statistics(
+        null_count: usize,
+        distinct_count: usize,
+        min_value: Option<ScalarValue>,
+        max_value: Option<ScalarValue>,
+        sum_value: Option<ScalarValue>,
+    ) -> ColumnStatistics {
+        ColumnStatistics {
+            null_count: datafusion_common::stats::Precision::Exact(null_count),
+            distinct_count: datafusion_common::stats::Precision::Exact(distinct_count),
+            min_value: min_value.map_or_else(
+                || datafusion_common::stats::Precision::Absent,
+                datafusion_common::stats::Precision::Exact,
+            ),
+            max_value: max_value.map_or_else(
+                || datafusion_common::stats::Precision::Absent,
+                datafusion_common::stats::Precision::Exact,
+            ),
+            sum_value: sum_value.map_or_else(
+                || datafusion_common::stats::Precision::Absent,
+                datafusion_common::stats::Precision::Exact,
+            ),
+        }
+    }
+
+    #[test]
+    fn test_map_column_statistics_basic_struct() -> Result<()> {
+        let file_schema = create_basic_nested_schema();
+        let table_schema = create_deep_nested_schema();
+
+        let adapter = DefaultSchemaAdapterFactory
+            .create(Arc::clone(&table_schema), Arc::clone(&table_schema));
+        let (mapper, _) = adapter.map_schema(file_schema.as_ref())?;
+
+        let file_stats = vec![create_test_column_statistics(
+            5,
+            100,
+            Some(ScalarValue::Utf8(Some("min_value".to_string()))),
+            Some(ScalarValue::Utf8(Some("max_value".to_string()))),
+            Some(ScalarValue::Utf8(Some("sum_value".to_string()))),
+        )];
+
+        let table_stats = mapper.map_column_statistics(&file_stats)?;
+
+        assert_eq!(table_stats.len(), 1);
+        verify_column_statistics(
+            &table_stats[0],
+            Some(5),
+            Some(100),
+            Some(ScalarValue::Utf8(Some("min_value".to_string()))),
+            Some(ScalarValue::Utf8(Some("max_value".to_string()))),
+            Some(ScalarValue::Utf8(Some("sum_value".to_string()))),
+        );
+
+        Ok(())
+    }
+
+    #[test]
+    fn test_map_column_statistics_empty_struct() -> Result<()> {
+        let file_schema = create_basic_nested_schema();
+        let table_schema = create_deep_nested_schema();
+
+        let adapter = DefaultSchemaAdapterFactory
+            .create(Arc::clone(&table_schema), Arc::clone(&table_schema));
+        let (mapper, _) = adapter.map_schema(file_schema.as_ref())?;
+
+        let empty_stats = vec![];
+        let mapped_empty_stats = mapper.map_column_statistics(&empty_stats)?;
+
+        assert_eq!(mapped_empty_stats.len(), 1);
+        assert_eq!(mapped_empty_stats[0], ColumnStatistics::new_unknown());
+
+        Ok(())
+    }
+
+    #[test]
+    fn test_map_column_statistics_multiple_columns_struct() -> Result<()> {
+        let file_schema = Arc::new(Schema::new(vec![
+            Field::new("id", DataType::Int32, false),
+            Field::new(
+                "additionalInfo",
+                DataType::Struct(
+                    vec![
+                        Field::new("location", DataType::Utf8, true),
+                        Field::new(
+                            "timestamp_utc",
+                            DataType::Timestamp(
+                                TimeUnit::Millisecond,
+                                Some("UTC".into()),
+                            ),
+                            true,
+                        ),
+                    ]
+                    .into(),
+                ),
+                true,
+            ),
+        ]));
+
+        let table_schema = Arc::new(Schema::new(vec![
+            Field::new("id", DataType::Int32, false),
+            Field::new(
+                "additionalInfo",
+                DataType::Struct(
+                    vec![
+                        Field::new("location", DataType::Utf8, true),
+                        Field::new(
+                            "timestamp_utc",
+                            DataType::Timestamp(
+                                TimeUnit::Millisecond,
+                                Some("UTC".into()),
+                            ),
+                            true,
+                        ),
+                        Field::new(
+                            "reason",
+                            DataType::Struct(
+                                vec![Field::new("_level", DataType::Float64, true)]
+                                    .into(),
+                            ),
+                            true,
+                        ),
+                    ]
+                    .into(),
+                ),
+                true,
+            ),
+            Field::new("status", DataType::Utf8, true),
+        ]));
+
+        let adapter = DefaultSchemaAdapterFactory
+            .create(Arc::clone(&table_schema), Arc::clone(&table_schema));
+        let (mapper, _) = adapter.map_schema(file_schema.as_ref())?;
+
+        let file_stats = vec![
+            create_test_column_statistics(
+                0,
+                100,
+                Some(ScalarValue::Int32(Some(1))),
+                Some(ScalarValue::Int32(Some(100))),
+                Some(ScalarValue::Int32(Some(5100))),
+            ),
+            create_test_column_statistics(10, 50, None, None, None),
+        ];
+
+        let table_stats = mapper.map_column_statistics(&file_stats)?;
+
+        assert_eq!(table_stats.len(), 3);
+        verify_column_statistics(
+            &table_stats[0],
+            Some(0),
+            Some(100),
+            Some(ScalarValue::Int32(Some(1))),
+            Some(ScalarValue::Int32(Some(100))),
+            Some(ScalarValue::Int32(Some(5100))),
+        );
+        verify_column_statistics(&table_stats[1], Some(10), Some(50), None, None, None);
+        assert_eq!(table_stats[2], ColumnStatistics::new_unknown());
+
+        Ok(())
     }
 }
